chore: introduce dep graph with meta [OSF-122]
In anticipation of adding new dep graph metadata (e.g. `target` and
`targetNameFromPlugin`), refactoring the data structure return from the
dep graph creation service.

Before, the service returned two parallel arrays of primitives instead
of a single array of well-structured data. This model was not only
confusing but also did not lent itself to adding more fields, hence this
set of changes was made.

This commit only brings structural changes, there is no difference in
terms of the behaviour.